Add jenis_surat column to Surat model and surats table

+++ app/Livewire/SuratTable.php
@@ -6,6 +6,7 @@ use App\Models\Surat;
 use Barryvdh\DomPDF\Facade\Pdf;
 use Illuminate\Support\Carbon;
 use Illuminate\Database\Eloquent\Builder;
+use Masmerise\Toaster\Toastable;
 use PowerComponents\LivewirePowerGrid\Button;
 use PowerComponents\LivewirePowerGrid\Column;
 use PowerComponents\LivewirePowerGrid\Detail;
@@ -21,37 +22,38 @@ use PowerComponents\LivewirePowerGrid\Traits\WithExport;
 final class SuratTable extends PowerGridComponent
 {
     use WithExport;
+    use Toastable;
 
     public function setUp(): array
     {
         $this->showCheckBox();
 
-        return [
-            Exportable::make('export')
-                ->striped()
-                ->type(Exportable::TYPE_XLS, Exportable::TYPE_CSV),
+        $setUp = [
             Header::make()->showSearchInput(),
             Footer::make()
                 ->showPerPage()
                 ->showRecordCount(),
-                Detail::make()
-                ->showCollapseIcon()
-                ->view('details.surat-detail'),
         ];
+
+        if (auth()->user()->can('export')) {
+            $setUp[] = Exportable::make('export')
+                ->striped()
+                ->type(Exportable::TYPE_XLS, Exportable::TYPE_CSV);
+        }
+
+        return $setUp;
     }
 
     public function datasource(): Builder
     {
-        return Surat::query()
-            ->leftJoin('users', 'surat.user_id', '=', 'users.id')
-            ->select('surat.*', 'users.name as name','users.no_hp as no_hp','users.alamat as alamat')
-        ;
-
+        return Surat::query()->with(['user', 'desa']);
     }
 
     public function relationSearch(): array
     {
-        return ['users'=> ['name','no_hp','alamat']
+        return [
+            'user' => ['name'],
+            'desa' => ['nama_desa']
         ];
     }
 
@@ -59,52 +61,36 @@ final class SuratTable extends PowerGridComponent
     {
         return PowerGrid::fields()
             ->add('id')
+            ->add('user_id')
+            ->add('desa_id')
+            ->add('jenis_surat')
+            ->add('pengirim')
             ->add('perihal')
             ->add('tanggal_kegiatan')
             ->add('hari')
-            ->add('jam_kegiatan')
+            ->add('waktu')
             ->add('lokasi_kegiatan')
-            ->add('name')
-            ->add('no_hp')
-            ->add('alamat');
-        ;
+            ->add('status')
+            ->add('file_surat');
     }
 
     public function columns(): array
     {
         return [
-            Column::make('Nama','name')
+            Column::make('Nama', 'user.name')
                 ->searchable()
                 ->sortable(),
 
-            Column::make('No Hp','no_hp')   
+            Column::make('Jenis Surat', 'jenis_surat')
                 ->searchable()
                 ->sortable(),
 
-            Column::make('Alamat','alamat')
+            Column::make('Perihal', 'perihal')
                 ->searchable()
                 ->sortable(),
 
-            // Column::make('Perihal','perihal')
-            //     ->searchable()
-            //     ->sortable(),
-
-            // Column::make('Tanggal Kegiatan','tanggal_kegiatan')
-            //     ->searchable()
-            //     ->sortable(),
-
-            // Column::make('Hari','hari')
-            //     ->searchable()
-            //     ->sortable(),
-
-            // Column::make('Jam Kegiatan','jam_kegiatan')
-            //     ->searchable()
-            //     ->sortable(),
-
-            // Column::make('Lokasi Kegiatan','lokasi_kegiatan')
-            //     ->searchable()
-            //     ->sortable(),
-                
+            Column::make('Created At', 'created_at')
+                ->sortable(),
 
             Column::action('Action')
         ];
@@ -113,55 +99,72 @@ final class SuratTable extends PowerGridComponent
     public function filters(): array
     {
         return [
+            Filter::select('jenis_surat', 'jenis_surat')
+                ->dataSource(Surat::all()->pluck('jenis_surat')->unique())
+                ->optionLabel('jenis_surat')
+                ->optionValue('jenis_surat'),
+            Filter::select('status', 'status')
+                ->dataSource(Surat::all()->pluck('status')->unique())
+                ->optionLabel('status')
+                ->optionValue('status'),
         ];
-    }
-
-    #[\Livewire\Attributes\On('edit')]
-    public function edit($rowId): void
-    {
-        $this->js('alert('.$rowId.')');
     }
 
     public function actions(\App\Models\Surat $row): array
     {
-        return [
-            Button::add('edit')
+        $actions = [];
+
+        if (auth()->user()->can('update')) {
+            $actions[] = Button::add('edit')
                 ->slot('<svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5" stroke="currentColor" class="w-5 h-5">
                 <path stroke-linecap="round" stroke-linejoin="round" d="m16.862 4.487 1.687-1.688a1.875 1.875 0 1 1 2.652 2.652L10.582 16.07a4.5 4.5 0 0 1-1.897 1.13L6 18l.8-2.685a4.5 4.5 0 0 1 1.13-1.897l8.932-8.931Zm0 0L19.5 7.125M18 14v4.75A2.25 2.25 0 0 1 15.75 21H5.25A2.25 2.25 0 0 1 3 18.75V8.25A2.25 2.25 0 0 1 5.25 6H10" />
                 </svg>
                 ')
                 ->id()
                 ->class('pg-btn-white dark:ring-pg-primary-600 dark:border-pg-primary-600 dark:hover:bg-pg-primary-700 dark:ring-offset-pg-primary-800 dark:text-pg-primary-300 dark:bg-pg-primary-700')
-                ->openModal('surat-form', ['rowId' => $row->id]),
-            Button::add('delete')
+                ->openModal('surat-form', ['rowId' => $row->id]);
+        }
+
+        if (auth()->user()->can('delete')) {
+            $actions[] = Button::add('delete')
                 ->slot('<svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5" stroke="currentColor" class="w-5 h-5">
                 <path stroke-linecap="round" stroke-linejoin="round" d="m14.74 9-.346 9m-4.788 0L9.26 9m9.968-3.21c.342.052.682.107 1.022.166m-1.022-.165L18.16 19.673a2.25 2.25 0 0 1-2.244 2.077H8.084a2.25 2.25 0 0 1-2.244-2.077L4.772 5.79m14.456 0a48.108 48.108 0 0 0-3.478-.397m-12 .562c.34-.059.68-.114 1.022-.165m0 0a48.11 48.11 0 0 1 3.478-.397m7.5 0v-.916c0-1.18-.91-2.164-2.09-2.201a51.964 51.964 0 0 0-3.32 0c-1.18.037-2.09 1.022-2.09 2.201v.916m7.5 0a48.667 48.667 0 0 0-7.5 0" />
                 </svg>
                 ')
                 ->id()
                 ->class('pg-btn-white dark:ring-pg-primary-600 dark:border-pg-primary-600 dark:hover:bg-pg-primary-700 dark:ring-offset-pg-primary-800 dark:text-pg-primary-300 dark:bg-pg-primary-700')
-                ->dispatch('delete', ['rowId' => $row->id])
-        ];
+                ->dispatch('delete', ['rowId' => $row->id]);
+        }
+
+        return $actions;
     }
 
     public function header(): array
     {
-        return [
-            Button::add('add-surat')
+        $header = [];
+
+        if (auth()->user()->can('create')) {
+            $header[] = Button::add('add-surat')
                 ->slot(__('<svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5" stroke="currentColor" class="w-5 h-5">
                 <path stroke-linecap="round" stroke-linejoin="round" d="M12 4.5v15m7.5-7.5h-15" />
                 </svg>
                 '))
                 ->class('pg-btn-white dark:ring-pg-primary-600 dark:border-pg-primary-600 dark:hover:bg-pg-primary-700 dark:ring-offset-pg-primary-800 dark:text-pg-primary-300 dark:bg-pg-primary-700 w-full')
-                ->openModal('surat-form', []),
-            Button::add('export-pdf')
+                ->openModal('surat-form', []);
+        }
+
+        if (auth()->user()->can('export')) {
+            $header[] =
+                Button::add('export-pdf')
                 ->slot(__('<svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5" stroke="currentColor" class="w-5 h-5">
                 <path stroke-linecap="round" stroke-linejoin="round" d="M6.72 13.829c-.24.03-.48.062-.72.096m.72-.096a42.415 42.415 0 0 1 10.56 0m-10.56 0L6.34 18m10.94-4.171c.24.03.48.062.72.096m-.72-.096L17.66 18m0 0 .229 2.523a1.125 1.125 0 0 1-1.12 1.227H7.231c-.662 0-1.18-.568-1.12-1.227L6.34 18m11.318 0h1.091A2.25 2.25 0 0 0 21 15.75V9.456c0-1.081-.768-2.015-1.837-2.175a48.055 48.055 0 0 0-1.913-.247M6.34 18H5.25A2.25 2.25 0 0 1 3 15.75V9.456c0-1.081.768-2.015 1.837-2.175a48.041 48.041 0 0 1 1.913-.247m10.5 0a48.536 48.536 0 0 0-10.5 0m10.5 0V3.375c0-.621-.504-1.125-1.125-1.125h-8.25c-.621 0-1.125.504-1.125 1.125v3.659M18 10.5h.008v.008H18V10.5Zm-3 0h.008v.008H15V10.5Z" />
                 </svg>
                 '))
                 ->class('pg-btn-white dark:ring-pg-primary-600 dark:border-pg-primary-600 dark:hover:bg-pg-primary-700 dark:ring-offset-pg-primary-800 dark:text-pg-primary-300 dark:bg-pg-primary-700 w-full')
-                ->dispatch('exportPdf', []),
-        ];
+                ->dispatch('exportPdf', []);
+        }
+
+        return $header;
     }
 
     public function gitlisteners()
@@ -172,46 +175,35 @@ final class SuratTable extends PowerGridComponent
                 'delete',
                 'exportPdf',
                 'edit',
-                'surat-updated' => '$refresh',
-            ] 
+                'suratUpdated' => '$refresh',
+            ]
         );
     }
-    
-     // Function to export PDF using DomPDF
-     public function exportPdf()
-     {
-         $path = public_path() . '/pdf';
-         // Mendapatkan datasource
-         $datasource = $this->datasource()->get();
-         // Membuat folder pdf jika belum ada
-         if (!file_exists($path)) {
-             mkdir($path, 0777, true);
-         }
-         // Membuat file pdf
-         $pdf = Pdf::loadView('pdf.surat', ['datasource' => $datasource]);
-         // Menyimpan file pdf ke folder pdf
-         $pdf->save($path . '/surat.pdf');
-         // Menampilkan file pdf
-         return response()->download($path . '/surat.pdf');
-     }
- 
-     // Function to delete data
-     public function delete($rowId)
-     {
-         $surat = Surat::findOrFail($rowId);
-         // Detach all associated users
-         $surat->user()->detach();
-         $surat->delete();
-     }
-    /*
-    public function actionRules($row): array
+
+    // Function to export PDF using DomPDF
+    public function exportPdf()
     {
-       return [
-            // Hide button edit for ID 1
-            Rule::button('edit')
-                ->when(fn($row) => $row->id === 1)
-                ->hide(),
-        ];
+        $path = public_path() . '/pdf';
+        // Mendapatkan datasource
+        $datasource = $this->datasource()->get();
+        // Membuat folder pdf jika belum ada
+        if (!file_exists($path)) {
+            mkdir($path, 0777, true);
+        }
+        // Membuat file pdf
+        $pdf = Pdf::loadView('pdf.surat', ['datasource' => $datasource]);
+        // Menyimpan file pdf ke folder pdf
+        $pdf->save($path . '/surat.pdf');
+        // Menampilkan file pdf
+        return response()->download($path . '/surat.pdf');
     }
-    */
+
+    // Function to delete data
+    public function delete($rowId)
+    {
+        $surat = Surat::findOrFail($rowId);
+        $surat->delete();
+
+        $this->toast('success', 'Surat berhasil dihapus');
+    }
 }
